+++ crates/gateway/src/main.rs
@@ -7,8 +7,8 @@ use axum::{
 };
 use bf_common::tracing::setup_tracing;
 use blockfrost_gateway::{
-    api, blockfrost, config, db, hydra_server_bridge, hydra_server_platform, load_balancer,
-    middlewares, rate_limit, sdk_bridge_ws,
+    api, blockfrost, config, db, health_monitor, hydra_server_bridge, hydra_server_platform,
+    load_balancer, middlewares, rate_limit, sdk_bridge_ws,
 };
 use clap::Parser;
 use colored::Colorize;
@@ -42,6 +42,16 @@ async fn main() -> Result<()> {
     )
     .await;
     let blockfrost_api = blockfrost::BlockfrostAPI::new(&config.blockfrost.project_id);
+    let health_monitor =
+        health_monitor::HealthMonitor::spawn(pool.clone(), blockfrost_api.clone()).await;
+
+    // Fail fast on startup problems
+    let initial_health = health_monitor.current_status().await;
+    if !initial_health.healthy {
+        tracing::error!("Refusing to start unhealthy: {}", initial_health.details());
+        std::process::exit(1);
+    }
+
     let hydras_manager = if let Some(hydra_platform_config) = &config.hydra_platform {
         Some(
             hydra_server_platform::HydrasManager::new(
@@ -103,6 +113,7 @@ async fn main() -> Result<()> {
         .layer(Extension(load_balancer))
         .layer(Extension(config.clone()))
         .layer(Extension(pool))
+        .layer(Extension(health_monitor))
         .layer(Extension(blockfrost_api))
         .layer(Extension(register_rate_limiter))
         .layer(Extension(prometheus_handle));